// src/audio/clip/types.ts
export type ClipNodeState =
	| "initial"
	| "scheduled"
	| "started"
	| "stopped"
	| "paused"
	| "resumed"
	| "ended"
	| "disposed";